Update TimestreamSink.java
Fixed a bug that would cause RejectedRecords to never get removed from `bufferedRecords`.

+++ integrations/flink_connector/src/main/java/com/amazonaws/services/timestream/TimestreamSink.java
@@ -106,7 +106,7 @@ public class TimestreamSink extends RichSinkFunction<TimestreamPoint> implements
 
                     LOG.warn("Discarding Malformed Record ->" + rejectedRecords.get(i).toString());
                     LOG.warn("Rejected Record Reason ->" + 	rejectedRecords.get(i).getReason());
-                    bufferedRecords.remove(rejectedRecords.get(i).getRecordIndex());
+                    bufferedRecords.remove(rejectedRecords.get(i).getRecordIndex().intValue());
 
                 }
             }   catch (Exception e) {
